sopotek_trading_ai_server/backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
httpx>=0.28,<1.0
aiosqlite>=0.21,<1.0
asyncpg>=0.30,<1.0
SQLAlchemy>=2.0,<3.0
PyJWT>=2.8,<3.0
python-multipart>=0.0.20,<1.0
aiokafka>=0.12,<1.0
pytest>=8.4,<9.0
email-validator>=2.2,<3.0
